+++ crates/ruma-signatures/CHANGELOG.md
@@ -22,7 +22,9 @@ Breaking changes:
 Improvements:
 
 - Add `verify_canonical_json_bytes()` as a low-level function to check the
-  signature of canonical JSON bytes
+  signature of canonical JSON bytes.
+- Add `check_pdu_sizes()` to check the size limits of a PDU according to the
+  Matrix specification.
 
 # 0.17.1